requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
requests
pillow
